Convert DigestStream into JS-backed stream
The original implementation of `DigestStream` used a WritableStreamSink
and the old "internal" streams implementation. This is unnecessary and
makes the implementation less efficient that it otherwise could be due
to the operations having to bounce between the kj event loop and the
JS isolate lock. This change converts `DigestStream` into a JS-backed
stream where the operations occur within the isolate lock, avoiding
the bounce and avoiding having to reacquire the isolate lock after
each chunk, etc. Also avoids the need for IoContext addObject / IoOwn
dereferencing. Implementation should be safer and more efficient.

Also allow DigestStream to accept string inputs. If a string is written
to a DigestStream, it will be converted to UTF-8 bytes and included in
the digest.

Also allow DigestStream to support `Symbol.dispose`.

Improves tests a bit also.