Add fetch_standard_url compat flag
Adds a compat flag that makes it so that the fetch('...') method
variant that takes a URL string, the `Request` object constructor,
and redirects will use the standard URL parser.

Unfortunately, this introduces a double parsing of the URL in order
to minimize changes to the existing implementation. It would be
nice to refactor things to avoid the double parsing at some point.